netbsd/arm64 signal delivery will trash g (R28), so we need to recover it from ucontext.
split netbsd signal handling for cgo and go.
This is done so that we can simplify the recover of g in the arm64
case:
cgo: from tpidr_el0
go: from an offset of ucontext

Use ucontext to restore context, rather than hoping that saving
registers on the stack works (it doesn't - some of the registers
are modified, like g).

fix kqueue, from freebsd/arm64 code.

Adjust the code of setNonblock. EOR seems wrong??